+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+##### 2.2.1 - 25 February 2015
+
+###### Backwards compatible bug fixes
+- #311 - Does not properly throw error in find() (like other adapters) when the item cannot be found
+
 ##### 2.2.0 - 24 February 2015
 
 ###### Backwards compatible API changes